fixed test function, added a parallel driver for MPI sorting, updated makefile

+++ serial_sorting/merge_sort.h
@@ -14,8 +14,8 @@ Note: This is from Jeff Martin's Homework 4.
 #define MERGE_SORT_SERIAL
 #include <stdlib.h>
 
-void Merge(int *arr, int first_index, int middle_index, int last_index);
-void MergeSort(int *arr, int index1, int index2);
+void Merge(double *arr, int first_index, int middle_index, int last_index);
+void MergeSort(double *arr, int index1, int index2);
 
 /// @brief Implements Merge as described on p. 31 of CLRS edition 3. We assume that prior
 /// to calling this function, that the array is sorted between indices first_index and
@@ -25,14 +25,14 @@ void MergeSort(int *arr, int index1, int index2);
 /// @param first_index The first index
 /// @param middle_index The middle index
 /// @param last_index The last index
-void Merge(int *arr, int first_index, int middle_index, int last_index) {
+void Merge(double *arr, int first_index, int middle_index, int last_index) {
     int i, j, k;
 
     /* Make copies of the left and right sub arrays */
     int len1 = middle_index - first_index + 1;
     int len2 = last_index - middle_index;
-    int *left = (int*)malloc(len1 * sizeof(int));
-    int *right = (int*)malloc(len2 * sizeof(int));
+    double *left = (double*)malloc(len1 * sizeof(double));
+    double *right = (double*)malloc(len2 * sizeof(double));
     for (i = 0; i < len1; i++)
         left[i] = arr[first_index + i];
     for (i = 0; i < len2; i++)
@@ -68,7 +68,7 @@ void Merge(int *arr, int first_index, int middle_index, int last_index) {
 /// @param arr An array
 /// @param index1 The first index of the sub array to be sorted
 /// @param index2 The last index of the sub array to be sorted
-void MergeSort(int *arr, int index1, int index2) {
+void MergeSort(double *arr, int index1, int index2) {
     if (index1 < index2) {
         int index_mid = (index1 + index2) / 2;
         MergeSort(arr, index1, index_mid);
